// Web/src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        ExtractedSource: {
            /** Content */
            content: string;
            /** Title */
            title: string;
            /** Url */
            url?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}